Refactor into package and add line drawing & PDF export
