+++ README.md
@@ -8,6 +8,7 @@ A desktop application for precise PDF-based takeoffs, purpose-built for electric
 - 📏 Measure and mark areas, lengths, and symbols
 - ✏️ Supports drawing lines, shapes, and custom symbols (like Libre Draw)
 - 💾 Save and reload annotations (doodles, highlights, symbols)
+- 💾 Save annotated PDFs and "Save As" for printing later
 - 📤 Export annotated PDF for printing or sharing
 - 📊 Integrate with Excel (via OpenPyXL) for material/labor tracking
