+++ takeofftool/app.py
@@ -1,0 +1,15 @@
+from PyQt5 import QtWidgets
+import sys
+
+from .main_window import MainWindow
+
+
+def main():
+    app = QtWidgets.QApplication(sys.argv)
+    window = MainWindow()
+    window.show()
+    sys.exit(app.exec_())
+
+
+if __name__ == "__main__":
+    main()
